class-29/demo/template_site/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-appconf==1.0.5
django-compressor==4.3
django-crispy-forms==1.14.0
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.4.3
